Updated aks deploy script to allow for more granular install of infrastructure pieces

+++ scripts/DeployAKS.ps1
@@ -19,32 +19,47 @@
 ################################
 # Run 						   #
 ################################
-#Variables to change
-$resourceGroupName = "MyResourceGroup"
+######Variables to change######
+$resourceGroupName = "MyAKS"
 $resourceGroupLocation = "eastus2"
-$domainNameResourceGroup = "MyDomainResourceGroup"
-$domainName = "MyDomainName.com"
-$hostName = "MyHostName"
-$azureClientId = ""
-$azureClientSecret = ""
-$azureTennantName = ""
-$adminAccountForPlatform = ""
 
-#variables that dont need to change
-$resourceGroupNameForNodes = $resourceGroupName + "Nodes" 
-$aksClusterName = "MyAKSCluster"
-$staticIpOutName = "AKSStaticOutIP"
-$staticIpInName = "AKSStaticInIP"
+###SQL###
+$createSqlServer = $true
+$sqlServerResourceGroup = $resourceGroupName
 $sqlServerName = $resourceGroupName + "sql"
 $sqlDatabaseName = "Profisee"
 $sqlUserName = "serveradmin"
 $sqlPassword = "P@ssw0rd33#$"
-$clusterVmSizeForLinux = "Standard_B2s"
-$clusterVmSizeForWindows = "Standard_B4ms"
-$kubernetesVersion = "1.16.7"
+
+###DNS###
+$createDNSRecord = $true
+$domainNameResourceGroup = "domainRG"
+$domainName = "domain.com"
+$hostName = "aks"
+
+##File Repository###
+$createFileRepository = $true
 $storageAccountName = $resourceGroupName + "files" 
 $storageShareName = "files"
-$azureAuthorityUrl = "https://login.microsoftonline.com/" + $azureTennantName + ".onmicrosoft.com"
+
+###Azure AD App Registration###
+$createAppInAzureAD = $true
+$azureClientName = "My Identity Server"
+$azureClientId = ""
+$azureClientSecret = ""
+
+###Profisee platform###
+$adminAccountForPlatform = "MyEmail@domain.com"
+
+######Variables that dont need to change######
+$aksClusterName = "MyAKSCluster"
+$resourceGroupNameForNodes = "MC_" + $resourceGroupName + "_" + $aksClusterName + "_" + $resourceGroupLocation
+$staticIpOutName = "AKSStaticOutIP"
+$staticIpInName = "AKSStaticInIP"
+$clusterVmSizeForLinux = "Standard_B2s"
+$clusterVmSizeForWindows = "Standard_B4ms"
+$kubernetesVersion = "1.17.3"
+
 $externalDnsName = $hostName + "." + $domainName
 $externalDnsUrl = "https://" + $externalDnsName
 
@@ -54,7 +69,7 @@ az group delete --name $resourceGroupName --yes
 az group create --name $resourceGroupName --location $resourceGroupLocation
 
 #create aks cluster node pool (linox)
-az aks create --resource-group $resourceGroupName --name $aksClusterName --node-resource-group $resourceGroupNameForNodes --node-count 1 --enable-addons monitoring --kubernetes-version $kubernetesVersion --generate-ssh-keys --windows-admin-password P@ssw0rd1234! --windows-admin-username dude --vm-set-type VirtualMachineScaleSets --load-balancer-sku standard --network-plugin azure --node-vm-size $clusterVmSizeForLinux
+az aks create --resource-group $resourceGroupName --name $aksClusterName --node-count 1 --enable-addons monitoring --kubernetes-version $kubernetesVersion --generate-ssh-keys --windows-admin-password P@ssw0rd1234! --windows-admin-username winadmin --vm-set-type VirtualMachineScaleSets --load-balancer-sku standard --network-plugin azure --node-vm-size $clusterVmSizeForLinux
 
 #create static public ip for outbound usage - loadbalancer
 az network public-ip create --resource-group $resourceGroupNameForNodes --name $staticIpOutName --sku Standard --allocation-method static
@@ -75,36 +90,58 @@ az aks nodepool add --resource-group $resourceGroupName --cluster-name $aksClust
 az aks get-credentials --resource-group $resourceGroupName --name $aksClusterName --overwrite-existing
 
 #add dns record to domain
-az network dns record-set a delete -g $domainNameResourceGroup -z $domainName -n $hostName --yes
-az network dns record-set a add-record -g $domainNameResourceGroup -z $domainName -n $hostName -a $publicInIP --ttl 5
+if($createDNSRecord)
+{
+	az network dns record-set a delete -g $domainNameResourceGroup -z $domainName -n $hostName --yes
+	az network dns record-set a add-record -g $domainNameResourceGroup -z $domainName -n $hostName -a $publicInIP --ttl 5
+}
 
-#create sql server
-az sql server create --resource-group $resourceGroupName --name $sqlServerName --location $resourceGroupLocation -u $sqlUserName -p $sqlPassword
-$sqlServerFQDN = az sql server show --resource-group $resourceGroupName --name $sqlServerName --query fullyQualifiedDomainName --output tsv
+#create sql server - if needed
+if($createSqlServer)
+{
+	az sql server create --resource-group $resourceGroupName --name $sqlServerName --location $resourceGroupLocation -u $sqlUserName -p $sqlPassword
+}
 
+#get the sql server
+$sqlServerFQDN = az sql server show --resource-group $sqlServerResourceGroup --name $sqlServerName --query fullyQualifiedDomainName --output tsv
 
 #create firewall rule for sql server, ip of aks cluster and ip of where this script is run
-az sql server firewall-rule create --resource-group $resourceGroupName --server $sqlServerName.ToLower() --name "aks node ip" --start-ip-address $publicOutIP --end-ip-address $publicOutIP
+az sql server firewall-rule create --resource-group $sqlServerResourceGroup --server $sqlServerName.ToLower() --name "aks node ip" --start-ip-address $publicOutIP --end-ip-address $publicOutIP
 $myIP = Invoke-RestMethod http://ipinfo.io/json | Select -exp ip
-az sql server firewall-rule create --resource-group $resourceGroupName --server $sqlServerName.ToLower() --name "scripter ip" --start-ip-address $myIP --end-ip-address $myIP
+az sql server firewall-rule create --resource-group $sqlServerResourceGroup --server $sqlServerName.ToLower() --name "scripter ip" --start-ip-address $myIP --end-ip-address $myIP
 
-#create storage account and file share - file repo for platform
-az storage account create --resource-group $resourceGroupName --name $storageAccountName.ToLower() --location $resourceGroupLocation
-$storageAccountKey = (az storage account keys list --resource-group $resourceGroupName --account-name $storageAccountName --query "[0].value")
-az storage share create --account-name $storageAccountName.ToLower() --account-key $storageAccountKey --name $storageShareName
+#create storage account and file share - file repo for platform - if needed
+if($createFileRepository)
+{
+	az storage account create --resource-group $resourceGroupName --name $storageAccountName.ToLower() --location $resourceGroupLocation
+	$storageAccountKey = (az storage account keys list --resource-group $resourceGroupName --account-name $storageAccountName --query "[0].value")
+	az storage share create --account-name $storageAccountName.ToLower() --account-key $storageAccountKey --name $storageShareName
+}
 $fileRepoUserName = "Azure\\" +  $storageAccountName.ToLower()
 $fileRepoPath = "\\\\" + $storageAccountName.ToLower() + ".file.core.windows.net\\" + $storageShareName
+
+#Azure AD Client registration
+$azureAppReplyUrl = $externalDnsUrl + "/profisee/auth/signin-microsoft"
+$azureTenantId = az account show --query tenantId --output tsv
+$azureAuthorityUrl = "https://login.microsoftonline.com/" + $azureTenantId
+if($createAppInAzureAD)
+{
+	az ad app create --display-name $azureClientName --reply-urls $azureAppReplyUrl
+	$azureClientId = az ad app list --filter "displayname eq '$azureClientName'" --query '[0].appId'
+}
+az ad app update --id $azureClientId --add --reply-Urls $azureAppReplyUrl
 
 #install nginx
 helm repo add stable https://kubernetes-charts.storage.googleapis.com/ 
 helm install nginx stable/nginx-ingress --values .\nginxValues.yaml --set controller.service.loadBalancerIP=$publicInIP
+#helm uninstall nginx stable/nginx-ingress
 
 #install profisee platform
 helm repo add profisee https://profisee.github.io
 helm install profiseeplatform2020r1 profisee/profisee-platform --values .\Values.yaml --set sqlServer.name=$sqlServerFQDN --set sqlServer.databaseName=$sqlDatabaseName --set sqlServer.userName=$sqlUserName --set sqlServer.password=$sqlPassword --set profiseeRunTime.fileRepository.userName=$fileRepoUserName --set profiseeRunTime.fileRepository.password=$storageAccountKey --set profiseeRunTime.fileRepository.location=$fileRepoPath --set profiseeRunTime.oidc.authority=$azureAuthorityUrl --set profiseeRunTime.oidc.clientId=$azureClientId --set profiseeRunTime.oidc.clientSecret=$azureClientSecret --set profiseeRunTime.adminAccount=$adminAccountForPlatform --set profiseeRunTime.externalDnsUrl=$externalDnsUrl --set profiseeRunTime.externalDnsName=$externalDnsName
 
 #check status and wait for "Pulling" to finish
-#kubectl describe pods
+#kubectl describe pod profisee-0
 
 #remote in and look
 #kubectl exec -it profisee-0 powershell
